Refactor AgreementPeriodsFieldArray and apply focus on startDate on mount, refs ERM-575
