+++ src/components/AgreementPeriodsFieldArray/AgreementPeriodsFieldArray.js
@@ -5,23 +5,11 @@ import { Field } from 'react-final-form';
 
 import {
   Button,
-  Col,
-  Datepicker,
   Headline,
-  Row,
-  TextArea,
 } from '@folio/stripes/components';
 
-import { composeValidators, EditCard, withKiwtFieldArray } from '@folio/stripes-erm-components';
-import { validators } from '../utilities';
-
-const multipleOpenEndedPeriods = (...rest) => (
-  validators.multipleOpenEnded(...rest, 'ui-agreements.errors.multipleOpenEndedPeriods')
-);
-
-const overlappingPeriods = (...rest) => (
-  validators.overlappingDates(...rest, 'ui-agreements.errors.overlappingPeriod')
-);
+import { EditCard, withKiwtFieldArray } from '@folio/stripes-erm-components';
+import AgreementPeriodField from './AgreementPeriodField';
 
 class AgreementPeriodsFieldArray extends React.Component {
   static propTypes = {
@@ -46,60 +34,12 @@ class AgreementPeriodsFieldArray extends React.Component {
         key={index}
         onDelete={index !== 0 ? () => this.props.onDeleteField(index, period) : undefined}
       >
-        <div>
-          <Row>
-            <Col xs={4}>
-              <Field
-                backendDateStandard="YYYY-MM-DD"
-                dateFormat="YYYY-MM-DD"
-                component={Datepicker}
-                id={`period-start-date-${index}`}
-                label={<FormattedMessage id="ui-agreements.agreements.startDate" />}
-                name={`${name}[${index}].startDate`}
-                required
-                validate={composeValidators(
-                  validators.requiredStartDate,
-                  validators.dateOrder,
-                  overlappingPeriods,
-                )}
-              />
-            </Col>
-            <Col xs={4}>
-              <Field
-                backendDateStandard="YYYY-MM-DD"
-                dateFormat="YYYY-MM-DD"
-                component={Datepicker}
-                id={`period-end-date-${index}`}
-                label={<FormattedMessage id="ui-agreements.agreements.endDate" />}
-                name={`${name}[${index}].endDate`}
-                parse={v => v} // Lets us send an empty string instead of `undefined`
-                validate={composeValidators(
-                  validators.dateOrder,
-                  multipleOpenEndedPeriods,
-                  overlappingPeriods,
-                )}
-              />
-            </Col>
-            <Col xs={4}>
-              <Field
-                backendDateStandard="YYYY-MM-DD"
-                dateFormat="YYYY-MM-DD"
-                component={Datepicker}
-                id={`period-cancellation-deadline-${index}`}
-                label={<FormattedMessage id="ui-agreements.agreements.cancellationDeadline" />}
-                name={`${name}[${index}].cancellationDeadline`}
-                parse={v => v} // Lets us send an empty string instead of `undefined`
-              />
-            </Col>
-          </Row>
-          <Field
-            component={TextArea}
-            id={`period-note-${index}`}
-            label={<FormattedMessage id="ui-agreements.agreementPeriods.periodNote" />}
-            name={`${name}[${index}].note`}
-            parse={v => v} // Lets us send an empty string instead of `undefined`
-          />
-        </div>
+        <Field
+          component={AgreementPeriodField}
+          id={`ra-agreement-${index}`}
+          index={index}
+          name={`${name}[${index}]`}
+        />
       </EditCard>
     ));
   }
